+++ examples/basic2d.rs
@@ -1,20 +1,19 @@
 extern crate cgmath;
 extern crate rhusics;
-extern crate specs;
 extern crate shrev;
+extern crate specs;
 
 use cgmath::{Point2, Rad, Rotation2, Transform};
-use specs::{RunNow, World};
 use shrev::EventChannel;
+use specs::{RunNow, World};
 
-use rhusics::ecs::collide::prelude2d::{world_register, BasicCollisionSystem2, BodyPose2,
+use rhusics::ecs::collide::prelude2d::{register_collision, BasicCollisionSystem2, BodyPose2,
                                        BroadBruteForce2, CollisionMode, CollisionShape2,
-                                       CollisionStrategy, GJK2, Rectangle,
-                                       ContactEvent2};
+                                       CollisionStrategy, ContactEvent2, GJK2, Rectangle};
 
 pub fn main() {
     let mut world = World::new();
-    world_register::<BodyPose2, ()>(&mut world);
+    register_collision::<BodyPose2, ()>(&mut world);
 
     let mut reader_1 = world
         .write_resource::<EventChannel<ContactEvent2>>()
