+++ src/ecs/collide/prelude2d.rs
@@ -6,13 +6,11 @@ pub use collision::primitive::{Circle, ConvexPolygon, Particle2, Rectangle};
 pub use collide::{CollisionMode, CollisionStrategy};
 pub use collide::prelude2d::*;
 
-use std::fmt::Debug;
-
 use cgmath::{Point2, Transform};
 use collision::dbvt::DynamicBoundingVolumeTree;
 use collision::primitive::Primitive2;
-use specs::{Component, Entity, World};
 use shrev::EventChannel;
+use specs::{Component, Entity, World};
 
 use {NextFrame, Real};
 use collide::ContactEvent;
@@ -56,7 +54,7 @@ pub type DynamicBoundingVolumeTree2 = DynamicBoundingVolumeTree<
     ContainerShapeWrapper<Entity, Primitive2<Real>>,
 >;
 
-/// Utility method for registering 2D components and resources with
+/// Utility method for registering 2D collision components and resources with
 /// [`specs::World`](https://docs.rs/specs/0.9.5/specs/struct.World.html).
 ///
 /// # Parameters
@@ -69,7 +67,7 @@ pub type DynamicBoundingVolumeTree2 = DynamicBoundingVolumeTree<
 /// - `T`: Transform type that implements
 ///        [`Transform`](https://docs.rs/cgmath/0.15.0/cgmath/trait.Transform.html).
 /// - `Y`: Shape type, see `Collider`
-pub fn world_register<T, Y>(world: &mut World)
+pub fn register_collision<T, Y>(world: &mut World)
 where
     T: Transform<Point2<Real>> + Component + Send + Sync + 'static,
     Y: Send + Sync + 'static,
@@ -78,29 +76,5 @@ where
     world.register::<NextFrame<T>>();
     world.register::<CollisionShape2<T, Y>>();
     world.add_resource(EventChannel::<ContactEvent2>::new());
-}
-
-/// Utility method for registering 2D components and resources with
-/// [`specs::World`](https://docs.rs/specs/0.9.5/specs/struct.World.html).
-///
-/// Will include components and resources needed for spatial sorting/collision detection.
-/// Will call [`world_register`](fn.world_register.html).
-///
-/// # Parameters
-///
-/// - `world`: The [world](https://docs.rs/specs/0.9.5/specs/struct.World.html)
-/// to register components/resources in.
-///
-/// # Type parameters
-///
-/// - `T`: Transform type that implements
-///        [`Transform`](https://docs.rs/cgmath/0.15.0/cgmath/trait.Transform.html).
-/// - `Y`: Shape type, see `Collider`
-pub fn world_register_with_spatial<T, Y>(mut world: &mut World)
-where
-    T: Transform<Point2<Real>> + Component + Clone + Debug + Send + Sync + 'static,
-    Y: Send + Sync + 'static,
-{
-    world_register::<T, Y>(&mut world);
     world.add_resource(DynamicBoundingVolumeTree2::new());
 }
