+++ src/ecs/physics/systems/mod.rs
@@ -1,5 +1,9 @@
 //! Physics systems
 
-pub use self::impulse_solver::*;
+pub use self::contact_resolution::ContactResolutionSystem;
+pub use self::impulse_solver::ImpulseSolverSystem;
+pub use self::next_frame::NextFrameSetupSystem;
 
 mod impulse_solver;
+mod contact_resolution;
+mod next_frame;
